refactor component directory structure
