+++ components/collection-object-contact-form/collection-object-contact-form.tsx
@@ -7,7 +7,7 @@ import { Button } from '../ui/button';
 import { Input } from '../ui/input';
 import { Label } from '../ui/label';
 
-export default function ObjectContactForm({ item, formId }) {
+export default function CollectionObjectContactForm({ item, formId }) {
   const dict = getDictionary();
   const [state, handleSubmit] = useForm(formId);
