+++ components/collection-object-image/image-viewer.tsx
