+++ components/collection-object-image/open-seadragon-viewer.tsx
