+++ components/collection-object/description-row.tsx
