+++ components/collection-object/geographical-description-row.tsx
