+++ components/search-card/content-card.tsx
