+++ components/search-card/palette-card.tsx
