+++ components/search/search-agg.tsx
@@ -6,7 +6,7 @@ import { getBooleanValue } from '@/util/various';
 import { ChevronsUpDown, Plus, X } from 'lucide-react';
 
 import type { AggOption } from '@/types/aggOption';
-import { useDebounce } from '@/components/forms/debounce';
+import { useDebounce } from '@/util/debounce';
 import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/checkbox';
 import {
