+++ components/term/term-card.tsx
